docs: Color revamp
