+++ app/ng-framework-modules-category/color/color-examples.component.ts
@@ -2,7 +2,7 @@ import { Component, ChangeDetectionStrategy } from "@angular/core";
 import { Link } from "./../../link";
 
 let menuLinks = [
-    new Link("Creating Colors", "/color/creating-colors"),
+    new Link("Usage", "/color/usage"),
 ];
 
 @Component({
