+++ app/ng-framework-modules-category/color/color-examples.module.ts
@@ -3,8 +3,7 @@ import { NativeScriptRouterModule } from "nativescript-angular/router";
 import { NativeScriptCommonModule } from "nativescript-angular/common";
 import { NativeScriptFormsModule } from "nativescript-angular/forms";
 import { ColorExamplesComponent } from "./color-examples.component";
-import { CreatingColorsExampleComponent } from "./creating-colors/creating-colors.component";
-import { HexPipe } from "./creating-colors/creating-colors.component";
+import { UsageComponent } from "./usage/usage.component";
 import { TitleAndNavButtonModule } from "../../directives/title-and-nav-button.module";
 
 export const routerConfig = [
@@ -13,9 +12,9 @@ export const routerConfig = [
         component: ColorExamplesComponent
     },
     {
-        path: "creating-colors",
-        component: CreatingColorsExampleComponent,
-        data: { title: "Creating colors" }
+        path: "usage",
+        component: UsageComponent,
+        data: { title: "Usage" }
     }
 ];
 
@@ -28,7 +27,7 @@ export const routerConfig = [
         NativeScriptFormsModule,
         NativeScriptRouterModule.forChild(routerConfig)
     ],
-    declarations: [HexPipe, ColorExamplesComponent, CreatingColorsExampleComponent]
+    declarations: [ColorExamplesComponent, UsageComponent]
 })
 
 export class ColorExamplesModule {
