--- app/ng-framework-modules-category/color/creating-colors/article.md
@@ -1,14 +1,0 @@
-Using color module
-<snippet id='button-text-html'/>
-
-Creating a Color from a hex value.
-<snippet id='creating-hex-color-code'/>
-
-Creating a Color from a short hex value.
-<snippet id='creating-short-hex-color-code'/>
-
-Creating a Color from four ARGB values (alpha, red, green, blue)
-<snippet id='creating-argb-color'/>
-
-Creating a Color from a single ARGB value
-<snippet id='creating-single-argb-color'/>
