+++ app/ng-framework-modules-category/color/end.md
@@ -1,7 +1,12 @@
+## API References
 
-**API Reference for the** [Color Class](https://docs.nativescript.org/api-reference/classes/_color_.color.html)
+| Name     | Type    | 
+|----------|---------|
+| [tns-core-modules/color](https://docs.nativescript.org/api-reference/modules/_color_) | `Module` | 
+| [Color](https://docs.nativescript.org/api-reference/classes/_color_.color.html) | `Class` | 
+| [tns-core-modules/known-colors](https://docs.nativescript.org/api-reference/modules/_color_known_colors_) | `Module` | 
 
-**Native Component**
+## Native Component
 
 | Android               | iOS      |
 |:----------------------|:---------|
