+++ app/ng-framework-modules-category/color/usage/article.md
@@ -1,0 +1,1 @@
+<snippet id='creating-colors-code'/>
