+++ app/ng-framework-modules-category/color/usage/usage.component.html
@@ -1,0 +1,3 @@
+<GridLayout sdkExampleTitle sdkToggleNavButton>
+    <Label text="Refer to the code-behind file for Color module usage" textWrap="true" class="h3 text-center m-20"></Label>
+</GridLayout>
